fix store nav width and extract common functions to utils
