+++ src/js/libraryroot.custom.js
@@ -1,60 +1,4 @@
-const waitForElement = (selector, parent = document) => new Promise((resolve) => {
-    const el = parent.querySelector(selector);
-    if (el) {
-        resolve(el);
-    }
-
-    const observer = new MutationObserver(() => {
-        const el = parent.querySelector(selector);
-        if (!el) {
-            return;
-        }
-
-        resolve(el);
-        observer.disconnect();
-    });
-
-    observer.observe(document.body, {
-        subtree: true,
-        childList: true,
-    });
-});
-
-
-
-
-// Container/Element Tracker Utility
-// Tracks DOM element references and adds basic lifecycle hooks
-const createContainerTracker = (selector, { onCleanup, onSetup, onUpdate } = {}) => {
-    let container = null;
-
-    const rebind = () => {
-        const nextContainer = document.querySelector(selector);
-
-        if (nextContainer === container) {
-            return;
-        }
-
-        container = nextContainer;
-
-        if (!container) {
-            onCleanup?.();
-            return;
-        }
-
-        onSetup?.();
-        onUpdate?.();
-    };
-
-    return {
-        rebind,
-        getContainer: () => container,
-        setContainer: (newContainer) => { container = newContainer; },
-    };
-};
-
-
-
+import { waitForElement, createContainerTracker } from './utils.js';
 
 // Create Loading Screen
 const createLoadingDiv = () => {
@@ -98,13 +42,15 @@ function syncWidthIfTargetHidden() {
 
     const setWidth = () => {
         const width = sourceEl.style.width;
+        const scrollbarWidth = '16px';
         if (width) {
-            targetEl.style.width = width;
+            targetEl.style.flex = `0 0 calc(${width} + ${scrollbarWidth})`;
         }
     };
 
     const removeWidth = () => {
         targetEl.style.removeProperty('width');
+        targetEl.style.removeProperty('flex');
     };
 
     const handleTargetDisplayChange = () => {
@@ -371,7 +317,7 @@ syncUserpanelWidth();
 // Create userpanel button container and move buttons
 (async () => {
     var gameListSidebar = await waitForElement('._3x1HklzyDs4TEjACrRO2tB');
-    
+
     // Userpanel
     var userPanel = await waitForElement('._3cykd-VfN_xBxf3Qxriccm._1-9sir4j_KQiMqdkZjQN0u');
     var friendButton = await waitForElement('._1TdaAqMFadi0UTqilrkelR');
