+++ src/js/webkit.js
@@ -1,25 +1,9 @@
 import './webkit/community/friendsThatPlayed.js';
+import { waitForElement } from './utils.js';
 
 (async function() {
     // should work only on search page
     if (!window.location.pathname.startsWith('/search')) return;
-
-    const waitForElement = (selector, parent = document) => new Promise((resolve) => {
-        const el = parent.querySelector(selector);
-        if (el) resolve(el);
-
-        const observer = new MutationObserver(() => {
-            const el = parent.querySelector(selector);
-            if (!el) return;
-            resolve(el);
-            observer.disconnect();
-        });
-
-        observer.observe(document.body, {
-            subtree: true,
-            childList: true,
-        });
-    });
 
     const containerParent = await waitForElement('.searchbar');
     const searchResultsRows = await waitForElement('#search_resultsRows');
